polykybd: extract shared keymap logic to keyboard-level poly_keymap.c
The corne42 variant's keymap.c had drifted into a stale fork of split72's:
it was frozen at ~58 languages (the cog markers had been stripped, so every
cog-driven language addition only updated split72 — NUM_LANG is now 156), and
it was missing split72's render_key wide-glyph layout fixes, the fw-staging
firmware-update state machine, the USER_SYNC_FW_UP_* registrations, and the
QK_REBOOT slave-reboot sync.

Fix the drift structurally: move all shared behaviour (rendering, HID/overlay
handling, language selection, idle/suspend, split sync, firmware-update state
machine) out of each variant's keymaps/default/keymap.c into a single
keyboard-level poly_keymap.c, compiled for every variant via the keyboard
rules.mk. Each variant's keymap.c now carries ONLY its data: keymaps[],
encoder_map[], and (RGB variants only) g_led_config.

Variant differences resolve at compile time:
  * polykybd.h includes the active variant header via -DKEYBOARD_*, so
    QMK_KEYBOARD_H reaches struct display_info + the BITMASK* macros.
  * POLY_DISP_ROW_0/3 and POLY_SPLASH_* are per-variant header macros.
  * RGB_MATRIX_ENABLE / POINTING_DEVICE_ENABLE guard the RGB and Cirque
    paths (the few RGB calls split72 left unguarded are now guarded).

corne42 gains full parity automatically (156 languages, render fixes, split
firmware-update, reboot sync); +9.3 KB .text, well within the 2 MB budget.

split72 is verified UNCHANGED: identical .text/.data/.bss sizes and, per a
per-function normalized-disassembly comparison, all 1251 functions have an
identical instruction sequence (the raw .text differs only by relocated
addresses from code moving to a new object file). The two variants can no
longer drift apart.

cog blocks (the language tables) moved with the logic; run_cog.sh and the
in-file xlsx path are updated to target poly_keymap.c (verified idempotent).
