+++ keyboards/handwired/polykybd/polykybd.h
@@ -16,4 +16,15 @@
 
 #pragma once
 
+// Pull in the active variant's main header (struct display_info, the BITMASK*
+// macros and the POLY_DISP_ROW_* / POLY_SPLASH_* parameters) so any source that
+// reaches us through QMK_KEYBOARD_H — in particular the shared poly_keymap.c —
+// sees the variant definitions. QMK passes -DKEYBOARD_<filesafe-name>.
+#if defined(KEYBOARD_handwired_polykybd_split72)
+#    include "split72/split72.h"
+#elif defined(KEYBOARD_handwired_polykybd_split42)
+#    include "split42/split42.h"
+#elif defined(KEYBOARD_handwired_polykybd_corne42)
+#    include "corne42/corne42.h"
+#endif
