+++ keyboards/handwired/polykybd/run_cog.sh
@@ -2,7 +2,7 @@
 cog -r lang/lang_lut.c
 cog -r lang/lang_lut.h
 cog -r lang/named_glyphs.h
-cog -r split72/keymaps/default/keymap.c
+cog -r poly_keymap.c
 cog -r keycode_helper.h
 cog -r uni.h
 cog -r hid_com.c
